+++ src/main/java/alfred/exception/InvalidDateException.java
@@ -1,4 +1,4 @@
-package alfred;
+package alfred.exception;
 
 public class InvalidDateException extends AlfredException {
